feat: retrieve & tested

+++ src/pelit/route.py
@@ -1,10 +1,11 @@
-from flask import Blueprint, request, Response, jsonify
+from flask import Blueprint, request, Response, jsonify, send_file
 from typing import Any, Optional
 import os
 import secrets
 from pathlib import Path
 import hashlib
 from pelit.plib.log import p_logger
+import re
 
 def create_route(cfg: dict[str, Any], logger: p_logger) -> Blueprint:
     """
@@ -73,11 +74,11 @@ def create_route(cfg: dict[str, Any], logger: p_logger) -> Blueprint:
         size = sum(f.stat().st_size for f in storage.rglob('*') if f.is_file()) / 1024 / 1024
         
         if 'max' in cfg['storage']:
-            if size > cfg['storage']['max']:
+            if size > cfg['storage']['max'] and cfg['storage']['max'] != 0:
                 return 2
 
         if 'warn' in cfg['storage']:
-            if size > cfg['storage']['warn']:
+            if size > cfg['storage']['warn'] and cfg['storage']['warn'] != 0:
                 return 1
         
         return 0
@@ -206,6 +207,13 @@ def create_route(cfg: dict[str, Any], logger: p_logger) -> Blueprint:
             删除结果，包含 JSON 格式的响应和状态码
         """
         info_head = f"{request.remote_addr} {request.method} {request.path}"
+        if not _authenticate():
+            logger.warn(f"{info_head} 401: 认证失败")
+            return jsonify({
+                "success": False,
+                "message": "认证失败"
+            }), 401
+        
         path = Path(cfg['storage']['path']) / directory / file
         try:
             os.remove(str(path))
@@ -229,7 +237,7 @@ def create_route(cfg: dict[str, Any], logger: p_logger) -> Blueprint:
                 "message": "创建目录失败"
             }), 502
 
-    @main_route.route('/list')
+    @main_route.route('/list', methods=['GET'])
     def list_root() -> tuple[Response, int]:
         """
         列出根目录下所有目录
@@ -238,6 +246,13 @@ def create_route(cfg: dict[str, Any], logger: p_logger) -> Blueprint:
             JSON 格式的列表和响应码
         """
         info_head = f"{request.remote_addr} {request.method} {request.path}"
+        if not _authenticate():
+            logger.warn(f"{info_head} 401: 认证失败")
+            return jsonify({
+                "success": False,
+                "message": "认证失败"
+            }), 401
+        
         try:
             resp: dict[str, Any] = {
                 "success": True,
@@ -255,7 +270,7 @@ def create_route(cfg: dict[str, Any], logger: p_logger) -> Blueprint:
                 "message": "列举失败"
             }), 502
         
-    @main_route.route('/list/<directory>')
+    @main_route.route('/list/<directory>', methods=['GET'])
     def list_dir(directory: str) -> tuple[Response, int]:
         """
         列出指定目录下所有目录
@@ -264,6 +279,13 @@ def create_route(cfg: dict[str, Any], logger: p_logger) -> Blueprint:
             JSON 格式的列表和响应码
         """
         info_head = f"{request.remote_addr} {request.method} {request.path}"
+        if not _authenticate():
+            logger.warn(f"{info_head} 401: 认证失败")
+            return jsonify({
+                "success": False,
+                "message": "认证失败"
+            }), 401
+        
         try:
             resp: dict[str, Any] = {
                 "success": True,
@@ -286,5 +308,59 @@ def create_route(cfg: dict[str, Any], logger: p_logger) -> Blueprint:
                 "success": False,
                 "message": "列举失败"
             }), 502
+
+    @main_route.route('/<directory>/<file>', methods=['GET'])
+    def retrieve(directory: str, file: str) -> tuple[Response, int]:
+        """
+        获取指定文件，包括反盗链支持
+        在生产环境中，此功能应当由服务器提供，Pelit 的服务仅供测试
+
+        Args:
+            directory: 文件所在目录
+            file: 文件名
+        
+        Returns:
+            获取的文件和状态码
+        """
+        info_head = f"{request.remote_addr} {request.method} {request.path}"
+        
+        # 验证 Referer 请求头
+        referer =  request.headers.get("Referer")
+        if 'hotlink_block' in cfg['network']:
+            if cfg['network']['hotlink_block']:
+                if not referer or 'hotlink_whitelist' not in cfg['network']:
+                    logger.info(f"{info_head} 403 反盗链阻止")
+                    return Response("禁止外链 Hotlinking forbidden"), 403
+                match = False
+                for rule in cfg['network']['hotlink_whitelist']:
+                    assert isinstance(rule, str)
+                    assert isinstance(referer, str)
+                    # 也允许直接匹配（忽略 . 在 RegEx 中的作用）
+                    if referer == rule:
+                        match = True
+                        break
+                    if re.fullmatch(rule, referer):
+                        match = True
+                        break
+                if not match:
+                    logger.info(f"{info_head} 403 反盗链阻止")
+                    return Response("禁止外链"), 403
+        
+        # 禁止访问隐藏的文件
+        if directory.startswith('.') or file.startswith('.'):
+            return Response("禁止访问"), 403
+
+        # 返回文件
+        f_path = Path(cfg['storage']['path']) / directory / file
+        if not f_path.exists():
+            return Response("未找到文件"), 404
+        try:
+            logger.info(f"{info_head} 200")
+            return send_file(str(f_path)), 200
+        except Exception as e:
+            logger.warn(f"{info_head} 502 发送文件失败")
+            logger.warn(f"这是一个内部错误，请检查配置")
+            logger.warn(f"{e}")
+            return Response("内部错误"), 502
 
     return main_route
